FIx test and remove interface.
