+++ versions.lock
@@ -25,6 +25,7 @@ com.palantir.conjure.java.runtime:keystores:5.10.0 (2 constraints: 4f1968db)
 com.palantir.refreshable:refreshable:1.0.0 (2 constraints: e51817b1)
 com.palantir.ri:resource-identifier:1.1.0 (2 constraints: ed14f4b5)
 com.palantir.safe-logging:preconditions:1.14.0 (14 constraints: 36e73ec9)
+com.palantir.safe-logging:preconditions-assertj:1.14.0 (1 constraints: 3805353b)
 com.palantir.safe-logging:safe-logging:1.14.0 (10 constraints: e79dad4b)
 com.palantir.safethreadlocalrandom:safe-thread-local-random:0.1.0 (1 constraints: 0305ee35)
 com.palantir.tokens:auth-tokens:3.6.2 (3 constraints: 9227a52e)
@@ -43,6 +44,7 @@ jakarta.ws.rs:jakarta.ws.rs-api:2.1.6 (2 constraints: fb144cb7)
 org.apache.httpcomponents.client5:httpclient5:5.0.3 (1 constraints: 0a050736)
 org.apache.httpcomponents.core5:httpcore5:5.0.3 (3 constraints: 992a51a8)
 org.apache.httpcomponents.core5:httpcore5-h2:5.0.2 (1 constraints: 3b13053c)
+org.assertj:assertj-core:3.19.0 (3 constraints: 5625e34d)
 org.checkerframework:checker-qual:3.5.0 (2 constraints: 161a1444)
 org.hdrhistogram:HdrHistogram:2.1.12 (1 constraints: 3e103aa2)
 org.immutables:value:2.8.8 (1 constraints: 14051536)
@@ -52,7 +54,6 @@ org.slf4j:slf4j-api:1.7.30 (11 constraints: 94a6984d)
 [Test dependencies]
 com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.11.1 (1 constraints: 3705333b)
 com.google.code.findbugs:annotations:3.0.1 (1 constraints: 9e0aafc3)
-com.palantir.safe-logging:preconditions-assertj:1.14.0 (1 constraints: 3805353b)
 com.palantir.tracing:tracing-test-utils:4.11.0 (1 constraints: 38053e3b)
 com.spotify.dataenum:dataenum:1.3.2 (1 constraints: e91058c1)
 com.squareup.okhttp3:mockwebserver:3.13.1 (1 constraints: 3a053f3b)
@@ -74,7 +75,6 @@ org.apache.logging.log4j:log4j-slf4j-impl:2.14.0 (1 constraints: 39053b3b)
 org.apache.pdfbox:fontbox:2.0.17 (1 constraints: 180b71d8)
 org.apache.pdfbox:pdfbox:2.0.17 (1 constraints: b40c5915)
 org.apiguardian:apiguardian-api:1.1.0 (6 constraints: 18697c5a)
-org.assertj:assertj-core:3.19.0 (3 constraints: 5625e34d)
 org.assertj:assertj-guava:3.3.0 (1 constraints: 08050336)
 org.awaitility:awaitility:4.0.3 (1 constraints: 09050236)
 org.hamcrest:hamcrest:2.2 (3 constraints: f41da570)
